Refactor code: Routes folder renamed to Config, moved passport code to /config/passport.js
