+++ config/auth.js
